Use a transparent 1x1 image for headerURL
This is a hack for an upstream limitation in the theme API

We can set `images.headerURL` which gives us a *non-tiling* image
anchored to the upper right. That's where the first  of the doubles
appears.

We can set `images.additional_backgrounds` which gives us images
that have customizable alignment and tiling.

If we don't set `headerURL`, we don't get the additional backgrounds at
all. So, here's a hack to set a transparent pixel for `headerURL`

Fixes #22