Add check to see if Themer tab already exists
Fixes #119. Currently, when the browserAction button is pressed, a new
Themer generator tab is created. As an enhancement, these changes check
the current window's tabs for a Themer tab and if found, sets the Themer
tab to be active / focused. If a Themer tab is not found, then one is
created like the original implementation.

In order to query the window's tabs, the `tabs` permission had to be added.