fix: Fix theme migration logic, follow-up to #819 (#825)
In #819, there is supposed to be a compatibility layer for older
versions of the add-on. However, it never ran because the condition
used `&` instead of `&&`. Fixed that here.

Besides that, `normalizeTheme` started to mutate its input parameters,
which the code is not prepared to handle. To avoid any unexpected side
effects, the `normalizeTheme` logic was changed so that input parameters
are not mutated. Similarly, the `postMessage` compat layer also clones
the input to avoid mutating inputs.

The `convertToBrowserTheme` method was also simplified, by moving the
`popup_text` and `popup` normalization to `normalizeTheme`.
The `tab_loading`/`tab_line` logic was made unconditional, because due
to the use of `normalizeTheme`, the `tab_loading` property is never set.