fix: Create and document xpi deployment strategy
There was an attempt to package XPIs for AMO in #523.

The `build:extension:amo` script was never used, and should not be used
because the generated XPI has ID `FirefoxColor-production@mozilla.com`.

The `package:amo` script was correct, but the generated XPI as not
published anywhere.

This commit introduces the `xpi` script, which generates two XPI files,
whose purpose is documented in README.md.
The generated XPI is intentionally independent of the branch where it is
built (by CircleCI), to make it easier to test add-ons on the staging or
production sites before signing and publishing the XPI.