fix: don't throw when color is void for colors with alpha
In #929, support for advanced colors got added, which included radio
buttons to toggle between the default (undefined) color and the color
from the color picker.

When the color picker doesn't have a default, the input would be
undefined. Since #944, that triggers an error because
`normalizeThemeColor` would fail if the input is void.

Both bugs together prevent the user from activating the color picker
via the radio input.

This patch fixes the issue by ensuring that the ThemeColorsEditor always
uses the (non-empty) color from the color picker. And the issue with
`normalizeThemeColor` has been fixed, together with a regression test.